Moved Module Intervals to Scripts
I think waybar keeps spawning instances so moving it to the script will ensure that only one
instance of the script is run at a time. It could just be that reloading sway just spawns instances
without killing the old scripts. Regardless, this makes waybar do one execution for sure and then
just waits on the script to give output